streamlit/requirements.txt
streamlit
streamlit-extras
python-dotenv>=1.0.0
authlib>=1.0.0
pyjwt>=2.8.0
requests>=2.31.0
supabase>=2.0.0
Pillow>=10.0.0
